remove ospf properties question
