+++ projects/question/src/main/java/org/batfish/question/ospfprocess/OspfProcessConfigurationQuestion.java
@@ -2,6 +2,7 @@ package org.batfish.question.ospfprocess;
 
 import static com.google.common.base.MoreObjects.firstNonNull;
 
+import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import javax.annotation.Nonnull;
@@ -20,12 +21,30 @@ public final class OspfProcessConfigurationQuestion extends Question {
   private static final String PROP_PROPERTIES = "properties";
 
   @Nullable private String _nodes;
+  @Nonnull private NodeSpecifier _nodeSpecifier;
   @Nonnull private OspfPropertySpecifier _properties;
 
-  public OspfProcessConfigurationQuestion(
+  @JsonCreator
+  private static OspfProcessConfigurationQuestion OspfProcessConfigurationQuestion(
       @JsonProperty(PROP_NODES) @Nullable String nodes,
       @JsonProperty(PROP_PROPERTIES) OspfPropertySpecifier propertySpec) {
+    return new OspfProcessConfigurationQuestion(
+        nodes,
+        SpecifierFactories.getNodeSpecifierOrDefault(nodes, AllNodesNodeSpecifier.INSTANCE),
+        firstNonNull(propertySpec, OspfPropertySpecifier.ALL));
+  }
+
+  public OspfProcessConfigurationQuestion(
+      @Nonnull NodeSpecifier nodeSpecifier, @Nonnull OspfPropertySpecifier propertySpec) {
+    this(null, nodeSpecifier, propertySpec);
+  }
+
+  private OspfProcessConfigurationQuestion(
+      @Nullable String nodes,
+      @Nonnull NodeSpecifier nodeSpecifier,
+      @Nonnull OspfPropertySpecifier propertySpec) {
     _nodes = nodes;
+    _nodeSpecifier = nodeSpecifier;
     _properties = firstNonNull(propertySpec, OspfPropertySpecifier.ALL);
   }
 
@@ -48,7 +67,7 @@ public final class OspfProcessConfigurationQuestion extends Question {
   @JsonIgnore
   @Nonnull
   NodeSpecifier getNodesSpecifier() {
-    return SpecifierFactories.getNodeSpecifierOrDefault(_nodes, AllNodesNodeSpecifier.INSTANCE);
+    return _nodeSpecifier;
   }
 
   @JsonProperty(PROP_PROPERTIES)
