--- questions/experimental/ospfProperties.json
@@ -1,29 +1,0 @@
-{
-    "class": "org.batfish.question.ospfproperties.OspfPropertiesQuestion",
-    "differential": false,
-    "nodes": "${nodes}",
-    "properties": "${properties}",
-    "instance": {
-        "description": "Returns configuration parameters for OSPF routing processes.",
-        "longDescription": "Returns all properties of all OSPF processes present in the network. ospfProcessConfiguration question should be used instead as this question is deprecated and will be removed in the future.",
-        "instanceName": "ospfProperties",
-        "tags": [
-            "configuration",
-            "ospf"
-        ],
-        "variables": {
-            "nodes": {
-                "description": "Include nodes matching this name or regex",
-                "type": "nodeSpec",
-                "optional": true,
-                "displayName": "Nodes"
-            },
-            "properties": {
-                "description": "Include properties matching this regex",
-                "type": "ospfPropertySpec",
-                "optional": true,
-                "displayName": "Properties"
-            }
-        }
-    }
-}
