--- tests/questions/experimental/ospfProperties.ref
@@ -1,32 +1,0 @@
-{
-  "class" : "org.batfish.question.ospfproperties.OspfPropertiesQuestion",
-  "nodes" : ".*",
-  "properties" : "maximum-.*",
-  "differential" : false,
-  "includeOneTableKeys" : true,
-  "instance" : {
-    "description" : "Returns configuration parameters for OSPF routing processes.",
-    "instanceName" : "qname",
-    "longDescription" : "Returns all properties of all OSPF processes present in the network. ospfProcessConfiguration question should be used instead as this question is deprecated and will be removed in the future.",
-    "tags" : [
-      "configuration",
-      "ospf"
-    ],
-    "variables" : {
-      "nodes" : {
-        "description" : "Include nodes matching this name or regex",
-        "displayName" : "Nodes",
-        "optional" : true,
-        "type" : "nodeSpec",
-        "value" : ".*"
-      },
-      "properties" : {
-        "description" : "Include properties matching this regex",
-        "displayName" : "Properties",
-        "optional" : true,
-        "type" : "ospfPropertySpec",
-        "value" : "maximum-.*"
-      }
-    }
-  }
-}
